Docx writer: better handling of tables in lists (#5947).
Previously the content of each list cell was indented when
the table belonged to a list item. This is obviously unwanted.
With this patch, a table inside a list item will be formatted
just like a table outside the list.

One desirable improvement would be to add some left indentation
to the entire table, so it visually appears to belong to the
list item. But I'm not sure how to do that.