BibTeX parser: fix handling of `%` in url field.
`%` does not function as a comment character inside `url`
(where URL-encoding is common).

Commit 6fb2973a582116cd515c6f7e68794cca22955511 mistakenly
took this reassignment of `%` to be a general feature of
braced (but not quoted) BibTeX fields.

This commit restores the correct behavior of `%` in braced fields
other than `url`, and corrects the behavior of `%` in `url`
when the value is quoted.

Closes #7678 (again).