Allow to control pandoc behavior with PANDOC_PROGRAM_NAME variable
If the variable is set, then this value will be used as the program name
to decide the behavior of the pandoc executable. Currently supported
values are `pandoc-server.cgi` for the CGI server, `pandoc-server` for
the HTTP(S) server, `pandoc-lua` for the Lua interpreter. Any other
value will trigger the usual behavior.