Update Zerobrane instructions to use Lua 5.4
Zerobrane IDE does not yet include Lua 5.4. I had tried to get 5.3 working with pandoc debugging but the debugger never triggered. Then I used a 5.4 install (I installed from homebrew) with mobdebug and luasocket installed. Setting this in Zerobrane's settings and configuring it in the IDE and remote debugging worked again. I've modified the instructions. Note I'm not a Lua person so I hope the details are clear...