MediaWiki writer: avoid extra section divs with repeated ids.
Closes #8383.

Previously when converting HTML with headings nested under
section elements, pandoc would produce HTML div tags and
a span with an identifier duplicating the one on the div.

It seems better just to omit the superstructure of section
divs, so that's what this commit does, yielding more
normal MediaWiki output.