clarify font selection for pdf -t ms
Previous versions of the documentation did not clarify much which
options of font selection are available for the generation of .pdf
via `-t ms`.  Nor was there a description how to extend the set of
of fonts provided by ghostscript/pdfgroff.  This commit equally
hints to Peter Schaffter's script install-fonts.sh to facilitate
greatly the later, too.