add required variables template error to Text.Pandoc.Writers.HTML
Allows an error to be thrown if a required variable is missing from the context.
This has only been applied to the HTML writer but could be added to all the
writers. I do not think this is a breaking change since existing templates do
not have required variables. It requires the new features in doctemplates pull
request https://github.com/jgm/doctemplates/pull/19