pandoc.cabal: remove pandoc.cabal, stack.cabal from extra-source-files.
The problem is that if these are in extra-source-files, then they
get put in the tarball, and then anyone trying to build the source
from an unpacked tarball will run into the problem that cabal.project
and stack.yaml refer to pandoc-server, pandoc-lua-engine, and
pandoc-cli, which aren't in the tarball.

See #8560.