Retain metadata in processing sections for chunked HTML.
Previously we suppressed metadata in all but the top page,
in order to prevent the title block from being printed
on every page. This prevented use of custom variables set
by metadata fields.

This commit moves to a better solution: a conditional
in the default template restricts the title block to the
top page.

Closes #8620.