HTML writer: Use different structure for epub footnotes.
Many EPUB readers are thrown off by pandoc's current footnote
output.  Both the ol and the fact that the footnote backlink is
at the end of the note seem to pose problems.

With this commit, we now create a list of aside (or div) elements,
instead of an ordered list. Each element begins with a note number
that is linked back to the note reference.  (So, the backlink occurs
at the beginning rather than the end.)

References:

- https://kindlegen.s3.amazonaws.com/AmazonKindlePublishingGuidelines.pdf
- http://kb.daisy.org/publishing/docs/html/notes.html

See #8672, closes #5583.

Thanks to @Porges and @lewer.