Fix apparent inaccuracy in MANUAL.txt
The section about pandoc's template syntax used to state that
spaces and tabs following the closing delimiter around variables
and control structures will be ignored. Some quick testing with
pandoc version 2.9.2.1 suggests that this is not the case. Spaces
and tabs _preceding_ the closing delimiter, however, are indeed
ignored. I therefore assume that this is simply a slight inaccuracy
in the manual; possibly produced by copying the sentence about the
opening delimiter and forgetting to adjust that word.