PandocMonad: allow nested data files
Data file fallbacks are now also searched in the directory that has the
same name as the file minus the extension. For example, a filter with
name `frob.lua` will be searched in `$DATADIR/filters/frob.lua` as well
as in `$DATADIR/filters/frob/frob.lua`.

This allows to place full git repositories in the data directory.

Closes: #6635