More fixes to 5e381e3.
These changes recognize that parseURI does not unescape the path.

Another change is that the canonical form of the path used as the
MediaBag key retains percent-encoding, if present; we only unescape
the string when writing to a file.

See #8918.

Some tests are needed before the issue can be closed.