Changing README to move formatting details to link
This changes the README of the pandoc project to be more readable for
new users, pushing the specific input and output formatting information
to a link, in the same fashion as the link to the installation page.
More friendly to new users and cordons off any additions to the
documentation to account for new input or output formats to a dedicated
page.