Fix default citeproc entry-spacing.
According to the CSL manual, the default entry spacing is 1.
We were treating it as 0.

T.P.Citeproc:  always include an entry-spacing attribute
in the Div if the bibliography element contains an entry-spacing
attribute (previously we omitted it when it was 0).

LaTeX writer: use entry spacing 1 if no entry-spacing
attribute is present.

Update tests.

See #9058.