Fix Latex theorem label issue as per #8872
In the current implementation , theorem labels in `theoremEnvironment` are determined by the `LastLabel` in the current state.
This approach works well when the `\label{label_name}` is placed at the end of the theorem body (, which is a standard place to put a label.)

However, when a label is placed at the beginning of the theorem (another common practice) and additional labels follow in the theorem body,
`theoremEnvironment` incorrectly picks the last label (e.g., `\label{item2}` in #8872).

This patch addresses the issue by extracting the label extraction independently of the `LastLabel` state.