Link pandoc-cli version to pandoc version.
Henceforth pandoc-cli's version will be synchronized with pandoc's, and
pandoc-cli will depend on an exact pandoc version.

This will avoid confusion by ensuring that `cabal install pandoc-cli-X.Y.Z`
installs pandoc version X.Y.Z.  It will make things more straightforward
for upstream packagers (see #9232).

Note also that the man pages included in this package are for a specific
pandoc version.

This scheme does not follow the Haskell PVP, but that should cause no
harm, because this package does not expose a library.