Typst reader: don't include metadata from document element.
The problem is that typst doesn't print this metadata; it is only
used in PDF properties. The title, authors, and so on are represented
in the typst document (and there is no standardized method like
LaTeX's `\maketitle`).  So if we parse this as metadata then converting
a typst document will likely lead to a double title.