test(Makefile): add validate-docx-golden-test2
Add a validation rule based on OOXML-Validator which uses the
Open-XML-SDK shipped and maintained as part of dotnet.

This catches a few more issues compared to the .xsd.

Unfortunately currently this executable always exits with 0 even when
validation failed. That can be fixed later by invoking it from a script.

Signed-off-by: Edwin Török <edwin@etorok.net>