ci(commit-validation): use git to walk commits
The PR event doesn't include the actual commits, just a count and a URL
to fetch it.
But we can checkout the entire git history, we don't have so many
commits that we need to optimize this just yet.
And then we might as well use `git` itself to walk the commit history.

(Using the remote commits URL would be useful only if we'd want to do a
shallow clone because we have too many commits)

Signed-off-by: Edwin Török <edwin@etorok.net>