fix(ci): explicitly choose a .Net version
Upstream OOXML-Validator has added support for building with .Net 8.0 in
commit 29af1086b5c78b6a213e70d2e1cb8da2a8e5b876.
But that means OOXML-Validator can now be built for both .Net 6.0 and 8.0,
so we need to pick which one we want when running, otherwise we get an
error:
```
Unable to run your project
Your project targets multiple frameworks.
Specify which framework to run using '--framework'.
```

Signed-off-by: Edwin Török <edwin@etorok.net>