LaTeX reader: better handling of colwidths.
Previously the parser just failed if the column width
specified in `p{}` wasn't a multiple of `\linewidth`.
This led to cases where content was skipped.
Now we treat these as ColWidthDefault and silently
parse the table.

A future improvement could be to guess relative column
widths from the dimensions specified, based on a default
line width.

Closes #9579.