poc: typst output properties
This is a proof of concept of output properties for Typst, allowing
Lua filters to translate CSS attributes/properties to Typst properties.

The Typst writer searches for attributes of the form typst:target:attr

Where target is "element" if the attribute should go to the element
or "text" if the content should be wrapped in a text element with this
attribute.

It is assumed that the value is raw Typst code suitable for insertion
as a property value, e.g. strings should be quoted and markup should
be bracketed.

The following cases are implemented:
- cell element
- table text
- block element
- span text

To be complete, each element which receives attributes would need to
process both kinds of attribute, and generate the text element if needed.