Typst output properties
This implements output of Typst properties from attributes of the
forms

* `typst:prop` - the value should be interpreted as raw Typst code
to set a parameter on the corresponding Typst element

* `typst:text:prop` - the value should be interpreted as raw Typst
code to set a parameter on a Typst `#set text()` rule, or text
element as appropriate.

This commit covers the following elements:
* span text
* div and div text
* table and table text
* td and td text

It would be possible to support more elements in the future.