Update MANUAL.txt to mention custom-style works with ODT
This commit https://github.com/jgm/pandoc/commit/bde3d7622ddd2cda7adabc309335bb2865c1d520 allowed `custom-style` to work with ODT but this is not mentioned in the manual. I tested what happens when the Style is not present and updated the text accordingly. I edited via Github web interface so hard wrapping is approximate. I hope the anchor to the reference doc works as expected (`#option--reference-doc` is the anchor as far as I could determine)...