RTF reader: internal refactoring.
We now use a HexChars token instead of HexChar; this allows
us to store sequences of escaped words, which will be important
for handling ANSI code pages that can have mulitbyte sequences.

Also, we add an optional parameter to ANSI for the code page.
So far this doesn't do anthing.

All prep for #9683.