Markdown reader: allow repeated labels in numbered example lists.
Previously if you tried to use the same label as an earlier
example list item, you'd get a new number, not the old one,
and references to the label would go to the second occurrence.

Now an existing label will be reused, and no new number will
be generated.  Caveat:  this only works reliably when the
re-used example list item occurs by itself in a list, or
occurs in a list of previously used example list items that
occur in exactly the same order as previously.