Lua: keep `lpeg` and `re` as "loaded" modules.
The modules `lpeg` and `re` are now treated as if they had been
loaded with `require`. Previously the modules were only assigned
to global values, but could be loaded again via `require`,
thereby allowing to use a system-wide installation. However, this
proved to be confusing.

The old behavior can be restored by adding the following lines to
the top of Lua scripts, or to the `init.lua` in the data dir.

    debug.registry()['_LOADED'].lpeg = nil
    debug.registry()['_LOADED'].re = nil