Update INSTALL.md
See #10069 -- the static linked pandoc on other OSes also
have a limitation to use binary lua packages. So added
this point at the top of the document. In addition
conda-forge is cross-platform way to install pandoc
and other tools, added to Windows, macOS and
Linux sections.

I wrapped at 75 characters, is that right?